Renamed AppSettings file and folder to SettingsScreen
